docker-infra Reorganized the packages.
