+++ src/main/java/tahafurkan/sandbox/usermanagement/Application.java
@@ -1,13 +1,13 @@
-package com.SpringBoot.SpringBootDatabaseProject;
+package tahafurkan.sandbox.usermanagement;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class SpringBootDatabaseProjectApplication {
+public class Application {
 
     public static void main(String[] args) {
-        SpringApplication.run(SpringBootDatabaseProjectApplication.class, args);
+        SpringApplication.run(Application.class, args);
     }
 
 }
