+++ src/main/java/tahafurkan/sandbox/usermanagement/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.SpringBoot.SpringBootDatabaseProject.repository;
+package tahafurkan.sandbox.usermanagement.repository;
 
-import com.SpringBoot.SpringBootDatabaseProject.entities.User;
+import tahafurkan.sandbox.usermanagement.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Integer> {
