+++ src/test/java/com/SpringBoot/SpringBootDatabaseProject/SpringBootDatabaseProjectApplicationTests.java
@@ -1,13 +1,9 @@
 package com.SpringBoot.SpringBootDatabaseProject;
 
-import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
 class SpringBootDatabaseProjectApplicationTests {
 
-	@Test
-	void contextLoads() {
-	}
 
 }
